기업정보extract/requirements.txt
# LangChain Framework & Core Components
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Store & Document Handling
faiss-cpu>=1.8.0
pymupdf>=1.24.6
pandas>=2.0.0
olefile>=0.47

# HWP File Parser
pyhwp>=0.1b15

# Agent Tools
ddgs>=5.3.1b1
wikipedia>=1.4.0
beautifulsoup4>=4.12.3
selenium>=4.21.0
requests>=2.32.0

# Web App Framework
streamlit>=1.35.0

# Environment Management
python-dotenv>=1.0.1

